refactor(lobby): move lobby list code to its own file
